预警执行日志  修改预警耗时

+++ facade/src/main/java/com/treefinance/saas/monitor/facade/domain/ro/AlarmExecuteLogRO.java
@@ -47,7 +47,7 @@ public class AlarmExecuteLogRO extends BaseRO{
     /**
      * 预警耗时
      */
-    private Integer costTime;
+    private Double costTime;
 
     /**
      * Info触发
